ui/pubsub_app/pyproject.toml
[project]
name = "pubsub"
version = "0.1.0"
description = ""
authors = [
    {name = "spidercatte",email = "catb_23@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standard] (>=0.34.3,<0.35.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "google-cloud-aiplatform (>=1.96.0,<2.0.0)",
    "google-adk (>=1.2.0,<2.0.0)",
    "google-cloud-secret-manager (>=2.23.3,<3.0.0)"
]

[tool.poetry]
packages = [{include = "pubsub_app", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
